third_party/blink/web_tests/external/wpt/webrtc/RTCDataChannel-binaryType.window-expected.txt

This is a testharness.js-based test.
FAIL Default binaryType value assert_equals: dc.binaryType should be 'blob' expected "blob" but got "arraybuffer"
FAIL Setting binaryType to 'blob' should succeed Failed to set the 'binaryType' property on 'RTCDataChannel': Blob support not implemented yet
PASS Setting binaryType to 'arraybuffer' should succeed
FAIL Setting binaryType to 'jellyfish' should be ignored Failed to set the 'binaryType' property on 'RTCDataChannel': Unknown binary type : jellyfish
FAIL Setting binaryType to 'arraybuffer ' should be ignored Failed to set the 'binaryType' property on 'RTCDataChannel': Unknown binary type : arraybuffer
FAIL Setting binaryType to '' should be ignored Failed to set the 'binaryType' property on 'RTCDataChannel': Unknown binary type :
FAIL Setting binaryType to 'null' should be ignored Failed to set the 'binaryType' property on 'RTCDataChannel': Unknown binary type : null
FAIL Setting binaryType to 'undefined' should be ignored Failed to set the 'binaryType' property on 'RTCDataChannel': Unknown binary type : undefined
FAIL Setting binaryType to '234' should be ignored Failed to set the 'binaryType' property on 'RTCDataChannel': Unknown binary type : 234
FAIL Setting binaryType to '54' should be ignored Failed to set the 'binaryType' property on 'RTCDataChannel': Unknown binary type : 54
Harness: the test ran to completion.
